Fix day formatting in clamav check script
The check that the clamav database has been updated correctly on a daily
basis has been failing consistently. The database itself has been
updated fine but the check doesn't work on single digit days.

For days with double digits the clamav log prints

`Mon Aug 31 18:08:43 2020`

However for days with a single digit the clamav log prints

`Wed Sep  9 17:30:04 2020`

There is a double space between the month and day :(

On days which only have a single digit we need to correctly format the
date in order to match the log line.