Add safebrowsing and copy cvd files
ClamAV has removed the ability to download the safebrowsing.cvd, as a result we need to ensure SafeBrowsing is equal to 'no'.
This is outlined in this issue: https://github.com/mko-x/docker-clamav/issues/103

ClamAV have limited the download of main.cvd, daily.cvd and bytecode.cvd files, as well as limiting the use of wget.
As a result, we need to copy these files from a source.
https://www.mail-archive.com/clamav-users@lists.clamav.net/msg49810.html