Replace scan with ping command (#71)
* Replace scan with ping command

clamdscan provides a ping command that responds PONG when the daemon is ready to accept scans. Otherwise it returns error.

On first try some days ago this didn't seem to work, always responding PONG even when the daemon was clearly not ready yet, but I've figured out the issue.

I was running the ping command against my running localhost daemon, not against the docker one. This has been validated by actually running the ping command (or the shell script) inside the container, instead of outside.

With this new ping command as it is not performing a virus scan, we can enable the `LogClean` directive so we can get a log every time a file has been scanned and help diagnose timeouts etc.

Doubled the `initialDelaySeconds` time because these pods really take a very long time to be even in a position to start probing them (I've seen it can take up to 4 or 5 mins).

* Delete branch from CI config