Improve probes and adds lifecycle preStop hook (#72)
* Tweaks to probe and preStop hook

This is for now just a test to proof some hypnotises.

Changed the probe from using the clamdscan ping command to use `tcpSocket `. Should have the same effect but is simpler. Reduce period from 15s to 5s.

Implemented a basic preStop hook as part of the pod lifecycle to ensure we kill gracefully clamd before allowing the pod to be killed to avoid any in-progress scans to finalise.

Maximum termination grace period if the preStop hang around for too long set to 90 seconds.

* Add a startup probe

Startup probe disables other probes until it is successful, then other probes take over and the startup probe stops.

Useful as the period can be increased (15s) in the startup probe and then reduced (5s) in the readiness.

* Prepare for deployment

* Increase grace period to 120s